requirements.txt
numpy
pandas
scikit-learn
tensorflow
torch
matplotlib
seaborn
jupyter
pytest
pyyaml